Password: Password Registration
Since qtkeychain cannot get all passwords from
the secure store, it is necessary to implement
some robust password registration system in all classes.
This can be implemented using static methods. There is only
one disadvantage, which is that I cannot find out the
registered passwords to DXC. But in this situation it is not
such a problem, because if the password to DXC is missing,
QLog will ask for this password.